Add PidExecutionManager and the ExecutionManagers package
This package contains different ExecutionManagers. An ExecutionManager
dictates for how long the monitor should run, usually monitoring some
info, such as PID, or by manually starting an application and stopping
it.

The PidExecutionManager is the first of these, and essentially
continuously monitors a PID until it detects that it is dead, signaling
that the application should terminate.

Articulation with the MiniPrexMonitor is still remaining